backend/requirements-api.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.25.1
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
redis==5.0.1
